Allow userAgent option to be a function
Some websites do not work with certain user agent settings. For example,
there is a website that tries to set cookies if it sees 'Mozilla' in the
user agent. If it cannot set the cookies, then the website refuses to
render any pages. That same website will render pages to the cURL user
agent without problem.